# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Invoice System API
  description: |
    REST API for the invoice reconciliation platform (CSV/XLSX accounting
    imports, reconciliation runs, notification emails, audit trail, users,
    company and application settings).

    ## Versioning

    * **Stable, documented namespace:** `/api/v1` (this document).
    * **Compatibility namespace:** `/api` resolves to the *same* handlers
      and remains supported so existing clients (including the bundled SPA
      frontend) keep working unchanged. No business logic is duplicated;
      both prefixes are registered against identical view functions.

    ## Response contract

    Every endpoint returns JSON. Success payloads use:

    ```json
    { "success": true, "message": "...", "data": { } }
    ```

    Error payloads always use the stable error envelope:

    ```json
    {
      "success": false,
      "message": "Human readable reason",
      "status": 400,
      "code": "BAD_REQUEST"
    }
    ```

    See the `Error` schema and the `codes` values in the components section.

    ## Correlation

    Every response carries an `X-Request-Id` header. A client-supplied value
    (alphanumeric/dot/dash, up to 64 chars) is echoed back; otherwise a
    generated UUID is returned. Use it to correlate client requests with
    server logs.

    ## Deprecation

    Superseded endpoints (marked `deprecated: true`) keep working unchanged,
...
    major version.

    ## Money and dates

...
  version: "1.0.0"
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: authLogin
      summary: Authenticate and issue access + refresh tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, example: admin@test.com }
                password: { type: string, format: password, example: password123 }
      responses:
        "200":
          description: Tokens issued.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessBody"
                  - type: object
                    properties:
                      data:
                        type: object
                        additionalProperties: true
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/InvalidCredentials" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /auth/me:
    get:
      tags: [Auth]
      operationId: authMe
      summary: Return the current authenticated user's profile.
      responses:
        "200":
          description: Profile payload.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessBody"
                  - type: object
                    properties:
                      data:
                        type: object
                        additionalProperties: true
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /users/:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List users for the admin's company.
      responses:
        "200":
          description: List of user objects.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessBody"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Users]
      operationId: createUser
      summary: Create a user (admin only).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username: { type: string }
                email: { type: string }
                password: { type: string, format: password }
                first_name: { type: string }
                last_name: { type: string }
                company_id: { type: integer, nullable: true }
                roles: { type: array, items: { type: string } }
                status: { type: string }
      responses:
        "201":
          description: User created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessBody"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /users/{user_id}:
    parameters:
      - { $ref: "#/components/parameters/UserId" }
    get:
      tags: [Users]
      operationId: getUser
      summary: Fetch a single user (admin only).
      responses:
        "200":
          description: User object.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessBody"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Users]
      operationId: updateUser
      summary: Update a user (admin only).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Updated user object.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessBody"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      tags: [Users]
      operationId: deleteUser
      summary: Delete a user (admin only).
      responses:
        "200":
          description: User deleted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token issued by POST /auth/login. Include as
        `Authorization: Bearer <token>`.
  parameters:
    UserId:
      in: path
      name: user_id
      required: true
      schema: { type: integer }
...
  headers:
    RequestId:
      description: Correlation ID echoed (or generated) for every response.
      schema: { type: string }
...
  schemas:
    SuccessBody:
      type: object
      required: [success]
      properties:
        success: { type: boolean, const: true }
        message:
          type: string
          description: Optional human-readable text (not required for the contract).
        data:
          description: >-
            Contract payload field. Present on endpoints that return data;
            omitted only for signal-only actions (logout, delete, password
            change). When present it is always an object or array, never a raw
            scalar.
          nullable: true
    Error:
      type: object
      required: [success, message, status, code]
      properties:
        success: { type: boolean, const: false }
        message: { type: string }
        status: { type: integer }
        code:
          type: string
          description: >-
            Stable machine-readable error code. `APP_ERROR` is a reserved
            internal fallback and must never be returned for the documented
            statuses below.
          enum:
            - BAD_REQUEST
            - INVALID_CREDENTIALS
            - UNAUTHORIZED
            - TOKEN_REQUIRED
            - TOKEN_EXPIRED
            - INVALID_TOKEN
            - TOKEN_REVOKED
            - FORBIDDEN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - VALIDATION_ERROR
            - RATE_LIMITED
            - DATABASE_ERROR
            - INTERNAL_ERROR
            - BAD_GATEWAY
            - SERVICE_UNAVAILABLE
            - MISSING_FIELD
...
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string }
        first_name: { type: string }
        last_name: { type: string }
        roles: { type: array, items: { type: string } }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
  responses:
...
    BadRequest:
      description: Invalid request.
      headers: *reqid_headers
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InvalidCredentials:
      description: Bad email/password.
      headers: *reqid_headers
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, expired, revoked or invalid token.
      headers: *reqid_headers
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but role insufficient.
      headers: *reqid_headers
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found.
      headers: *reqid_headers
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Rate limit exceeded.
      headers: *reqid_headers
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...